refactor new perspective test into two test files
